renaming to solt to handle multiple functions
