+++ build.gradle.kts
@@ -2,8 +2,12 @@ plugins {
     kotlin("multiplatform") version "1.4.10"
     kotlin("plugin.serialization") version "1.4.10"
 }
-group = "com.github.hjubb.solcsjw"
-version = "0.1.0"
+
+group = "com.github.hjubb"
+version = "0.4.0"
+
+val binomVersion = "0.1.22"
+val ktorVersion = "1.4.0"
 
 repositories {
     mavenCentral()
@@ -18,7 +22,7 @@ kotlin {
     macosX64 {
         binaries {
             executable {
-                entryPoint = "com.github.hjubb.solcsjw.main"
+                entryPoint = "com.github.hjubb.solt.main"
             }
         }
     }
@@ -26,7 +30,7 @@ kotlin {
     linuxX64 {
         binaries {
             executable {
-                entryPoint = "com.github.hjubb.solcsjw.main"
+                entryPoint = "com.github.hjubb.solt.main"
             }
         }
     }
@@ -34,7 +38,7 @@ kotlin {
     mingwX64 {
         binaries {
             executable {
-                entryPoint = "com.github.hjubb.solcsjw.main"
+                entryPoint = "com.github.hjubb.solt.main"
             }
         }
     }
@@ -43,8 +47,15 @@ kotlin {
             dependencies {
                 implementation("org.jetbrains.kotlinx:kotlinx-cli:0.3")
                 implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.0.0")
-                api("pw.binom.io:file:0.1.22")
-                api("pw.binom.io:core:0.1.22")
+                api("pw.binom.io:file:$binomVersion")
+                api("pw.binom.io:core:$binomVersion")
+                implementation("io.ktor:ktor-client-core:$ktorVersion")
+                implementation("io.ktor:ktor-client-curl:$ktorVersion")
+                implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.3.9-native-mt") {
+                    version {
+                        strictly("1.3.9-native-mt")
+                    }
+                }
             }
         }
         val commonTest by getting
